processors/kafka/sensordb-writer/src/requirements.txt
kafka-python 
psycopg2-binary